requirements.txt
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
yt-dlp==2023.11.16
openai-whisper==20231117
openai>=1.54.0
python-dotenv==1.0.1
tiktoken>=0.8.0
anyio==4.6.2.post1
